api_prod/requirements.txt
asgiref==3.5.0
Django==3.2.10
django-js-asset==2.0.0
django-mptt==0.13.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
psycopg2-binary==2.8.6
pytz==2022.1
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2022.1
gunicorn==20.0.4